requirements.txt
requests==2.31.0
python-dotenv==1.0.0
Flask==2.3.3
Werkzeug==2.3.7
